refactor: 重构请求日志entity结构

+++ platform-server/src/main/java/com/lmj/platformserver/constant/RequestResultConstant.java
@@ -2,7 +2,7 @@ package com.lmj.platformserver.constant;
 
 public class RequestResultConstant {
 
-    public static final int ERROR = -1;
-    public static final int FAIL = 0;
-    public static final int SUCCESS = 1;
+    public static final byte ERROR = -1;
+    public static final byte FAIL = 0;
+    public static final byte SUCCESS = 1;
 }
